requirements.txt
# Core
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Deep Learning
torch>=2.0.0
torch-geometric>=2.4.0

# Geospatial
geopandas>=0.14.0
rasterio>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Config
pyyaml>=6.0
tqdm>=4.65.0
